requirements_gpu.txt
nltk
numpy
tensorflow_gpu==1.1.0